--- extension/sdnotify/extension_integration_test.go
@@ -1,253 +1,0 @@
-// SPDX-FileCopyrightText: SAP SE or an SAP affiliate company and Gardener contributors
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package sdnotify
-
-import (
-	"context"
-	"strings"
-	"testing"
-	"time"
-
-	"github.com/moby/moby/api/types/container"
-	"github.com/moby/moby/api/types/mount"
-	"github.com/stretchr/testify/require"
-	"github.com/testcontainers/testcontainers-go"
-	"github.com/testcontainers/testcontainers-go/wait"
-)
-
-// TestSDNotify_SystemdIntegration runs otelcol as a Type=notify unit, and
-// asserts via systemctl that the unit reached active/running - which is only
-// possible if extension.Ready() successfully sent READY=1 over $NOTIFY_SOCKET.
-func TestSDNotify_SystemdIntegration(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
-	t.Cleanup(cancel)
-
-	req := testcontainers.ContainerRequest{
-		FromDockerfile: testcontainers.FromDockerfile{
-			Context:       "../..",
-			Dockerfile:    "extension/sdnotify/testdata/Dockerfile",
-			KeepImage:     false,
-			PrintBuildLog: true,
-		},
-		HostConfigModifier: func(hc *container.HostConfig) {
-			// Both kernel capabilities are required to run systemd in a container.
-			// https://github.com/systemd/systemd/blob/main/docs/CONTAINER_INTERFACE.md#what-you-shouldnt-do
-			hc.CapAdd = []string{"SYS_ADMIN", "MKNOD"}
-			// Should run without the default seccomp profile, because it denies mounting.
-			// https://docs.docker.com/engine/security/seccomp
-			hc.SecurityOpt = []string{"seccomp=unconfined"}
-			// Either pre-mount all cgroup hierarchies into the container, or leave that to systemd which will do so if they are missing.
-			// https://github.com/systemd/systemd/blob/main/docs/CONTAINER_INTERFACE.md#execution-environment
-			hc.CgroupnsMode = container.CgroupnsModeHost
-			hc.Mounts = []mount.Mount{
-				{
-					Type:   mount.TypeBind,
-					Source: "/sys/fs/cgroup",
-					Target: "/sys/fs/cgroup",
-				},
-			}
-		},
-		// Wait for systemd to finish booting AND for the otelcol.service to reach `active`.
-		WaitingFor: wait.ForExec([]string{
-			"systemctl", "is-active", "otelcol.service",
-		}).WithStartupTimeout(3 * time.Minute).
-			WithPollInterval(2 * time.Second).
-			WithExitCode(0),
-	}
-
-	ctr, err := testcontainers.GenericContainer(ctx, testcontainers.GenericContainerRequest{
-		ContainerRequest: req,
-		Started:          true,
-	})
-	require.NoError(t, err, "container failed to start; sdnotify integration prerequisites not met")
-	t.Cleanup(func() {
-		_ = ctr.Terminate(context.Background())
-	})
-
-	// Vaidate that otelcol.service is running & healthy.
-	show := execAndCollect(ctx, t, ctr,
-		"systemctl", "show", "otelcol.service",
-		"-p", "ActiveState", "-p", "SubState", "-p", "Result",
-	)
-	require.Contains(t, show, "ActiveState=active")
-	require.Contains(t, show, "SubState=running")
-	require.Contains(t, show, "Result=success")
-
-	// Confirm extension.Ready()'s log line landed in the journal.
-	journal := execAndCollect(ctx, t, ctr,
-		"journalctl", "-u", "otelcol.service", "--no-pager",
-	)
-	require.Contains(t, journal, "sent READY=1 to systemd",
-		"expected sdnotify READY=1 log line in journal, got:\n%s", journal,
-	)
-
-	// Stop the unit and verify a clean exit.
-	_ = execAndCollect(ctx, t, ctr, "systemctl", "stop", "otelcol.service")
-	stopped := execAndCollect(ctx, t, ctr,
-		"systemctl", "show", "otelcol.service",
-		"-p", "ActiveState", "-p", "Result",
-	)
-	require.Contains(t, stopped, "ActiveState=inactive")
-	require.Contains(t, stopped, "Result=success")
-}
-
-// TestSDNotify_SIGHUP_ReportsFatalErrorAndUnitFails asserts the SIGHUP path
-// end-to-end: with the default unit (Type=notify, Restart=no), sending
-// SIGHUP to the collector should cause the extension to send RELOADING=1,
-// then report a fatal component-status event, and the process should exit
-// with a non-zero status. systemd should mark the unit as failed -
-// distinguishing this from a clean stop.
-func TestSDNotify_SIGHUP_ReportsFatalErrorAndUnitFails(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
-	t.Cleanup(cancel)
-
-	ctr := startSystemdContainer(ctx, t,
-		[]string{"systemctl", "is-active", "otelcol.service"}, 0)
-
-	// Capture MainPID before signaling so we can assert the process actually died.
-	before := execAndCollect(ctx, t, ctr, "systemctl", "show", "otelcol.service", "-p", "MainPID")
-
-	_ = execAndCollect(ctx, t, ctr, "systemctl", "kill", "-s", "SIGHUP", "otelcol.service")
-
-	// With Restart=no the unit should settle into ActiveState=failed after
-	// the collector exits from the fatal-error event.
-	show := waitForUnitState(ctx, t, ctr, "failed", 30*time.Second)
-	require.Contains(t, show, "ActiveState=failed")
-	require.NotContains(t, show, "Result=success",
-		"expected non-success Result after SIGHUP-induced fatal error")
-
-	// Journal must show both halves of the extension's SIGHUP handling:
-	// the RELOADING=1 log line and the "reporting fatal error" log line.
-	journal := execAndCollect(ctx, t, ctr,
-		"journalctl", "-u", "otelcol.service", "--no-pager",
-	)
-	require.Contains(t, journal, "sent RELOADING=1 to systemd",
-		"expected RELOADING=1 log line after SIGHUP; journal:\n%s", journal)
-	require.Contains(t, journal, "reporting fatal error to trigger supervisor restart",
-		"expected fatal-error log line after SIGHUP; journal:\n%s", journal)
-
-	// MainPID should have changed to 0 (unit dead, not restarted).
-	after := execAndCollect(ctx, t, ctr, "systemctl", "show", "otelcol.service", "-p", "MainPID")
-	require.NotEqual(t, strings.TrimSpace(before), strings.TrimSpace(after),
-		"MainPID should change after SIGHUP-induced exit")
-	require.Contains(t, after, "MainPID=0")
-}
-
-// TestSDNotify_SIGHUP_WithRestartPolicy_RelaunchesUnit swaps in a
-// Restart=on-failure unit and asserts that SIGHUP causes a supervisor-driven
-// restart: the collector exits, systemd starts a fresh process, MainPID
-// changes, and the unit is back in active/running with NRestarts >= 1. This
-// is the documented "reload" flow: SIGHUP -> fatal exit -> supervisor
-// restart with the updated config.
-func TestSDNotify_SIGHUP_WithRestartPolicy_RelaunchesUnit(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
-	t.Cleanup(cancel)
-
-	// Start with the default unit (Restart=no) so we can boot to active
-	// deterministically, then swap in the restart-enabled unit and reload.
-	ctr := startSystemdContainer(ctx, t,
-		[]string{"systemctl", "is-active", "otelcol.service"}, 0)
-
-	installUnitFile(ctx, t, ctr, "testdata/otelcol-restart.service")
-	// Restart to pick up the new unit definition. The unit should return to
-	// active immediately since the config itself is unchanged.
-	_ = execAndCollect(ctx, t, ctr, "systemctl", "restart", "otelcol.service")
-	_ = waitForUnitState(ctx, t, ctr, "active", 30*time.Second)
-
-	beforePID := mainPID(ctx, t, ctr)
-	require.NotEmpty(t, beforePID, "expected non-empty MainPID before SIGHUP")
-
-	_ = execAndCollect(ctx, t, ctr, "systemctl", "kill", "-s", "SIGHUP", "otelcol.service")
-
-	// systemd should notice the failure, wait RestartSec, then start a new
-	// process. Wait for ActiveState=active with a new PID.
-	deadline := time.Now().Add(45 * time.Second)
-	var show, afterPID string
-	for time.Now().Before(deadline) {
-		show = execAndCollect(ctx, t, ctr,
-			"systemctl", "show", "otelcol.service",
-			"-p", "ActiveState", "-p", "SubState", "-p", "MainPID", "-p", "NRestarts",
-		)
-		afterPID = mainPID(ctx, t, ctr)
-		if strings.Contains(show, "ActiveState=active") &&
-			afterPID != "" && afterPID != "0" && afterPID != beforePID {
-			break
-		}
-		time.Sleep(500 * time.Millisecond)
-	}
-
-	require.Contains(t, show, "ActiveState=active",
-		"unit did not return to active after supervisor restart; show:\n%s", show)
-	require.NotEqual(t, beforePID, afterPID,
-		"MainPID should differ after supervisor restart; show:\n%s", show)
-	require.Contains(t, show, "NRestarts=1",
-		"expected NRestarts=1 after single SIGHUP-triggered restart; show:\n%s", show)
-
-	// The journal for the current (restarted) process should show a fresh
-	// READY=1 -- the second incarnation went through Ready() again.
-	journal := execAndCollect(ctx, t, ctr,
-		"journalctl", "-u", "otelcol.service", "--no-pager",
-	)
-	// Expect at least two READY=1 lines: the first boot and the restart.
-	readyCount := strings.Count(journal, "sent READY=1 to systemd")
-	require.GreaterOrEqual(t, readyCount, 2,
-		"expected >=2 READY=1 log lines (boot + restart), got %d; journal:\n%s", readyCount, journal)
-}
-
-// TestSDNotify_StartupFailure_LeavesUnitFailed asserts that when the
-// collector fails during Start() (before Ready() would fire), systemd's
-// Type=notify contract is upheld: the unit lands in failed state rather
-// than hanging on a READY=1 that never comes. This guards against
-// regressions where sdnotify might spuriously send READY=1 too early or
-// swallow the underlying startup error.
-func TestSDNotify_StartupFailure_LeavesUnitFailed(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
-	t.Cleanup(cancel)
-
-	// Boot with the default (working) config so systemd itself comes up;
-	// then swap in the broken config and restart the unit. We can't wait
-	// on `is-active` here because the unit will fail -- wait on systemd
-	// itself being up.
-	ctr := startSystemdContainer(ctx, t,
-		[]string{"systemctl", "is-system-running", "--wait"}, 0)
-
-	installConfigFile(ctx, t, ctr, "testdata/otelcol-bad-config.yaml")
-
-	// `systemctl restart` will block until start-limit is reached or the
-	// unit fails; with Restart=no it fails on the first attempt. Ignore
-	// exit code -- we assert on the resulting unit state.
-	_, _, _ = ctr.Exec(ctx, []string{"systemctl", "restart", "otelcol.service"})
-
-	show := waitForUnitState(ctx, t, ctr, "failed", 60*time.Second)
-	require.Contains(t, show, "ActiveState=failed",
-		"unit should be failed after startup with broken config; show:\n%s", show)
-
-	// The extension must NOT have logged a spurious READY=1 -- Ready()
-	// should never have been called on a failed startup.
-	journal := execAndCollect(ctx, t, ctr,
-		"journalctl", "-u", "otelcol.service", "--no-pager",
-	)
-	// Filter to just the failed incarnation (after the config swap).
-	// Simplest check: no successful READY=1 line since the last restart
-	// attempt would be a bug, but the first boot did send one. Assert on
-	// the presence of a collector-side error log instead.
-	require.True(t,
-		strings.Contains(journal, "error") || strings.Contains(journal, "failed"),
-		"expected error/failed log line from broken-config startup; journal:\n%s", journal)
-}
-
-// mainPID returns the MainPID= value from `systemctl show` for otelcol.service,
-// trimmed (just the number, or "0" when the unit isn't running).
-func mainPID(ctx context.Context, t *testing.T, ctr testcontainers.Container) string {
-	t.Helper()
-	out := execAndCollect(ctx, t, ctr, "systemctl", "show", "otelcol.service", "-p", "MainPID")
-	// Output looks like "MainPID=1234\n".
-	for _, line := range strings.Split(out, "\n") {
-		if v, ok := strings.CutPrefix(strings.TrimSpace(line), "MainPID="); ok {
-			return v
-		}
-	}
-	return ""
-}
