+++ extension/sdnotify/extension_test.go
@@ -6,43 +6,32 @@ package sdnotify
 
 import (
 	"context"
+	"io"
 	"net"
 	"os"
 	"path/filepath"
-	"sync"
+	"strings"
 	"syscall"
 	"testing"
 	"time"
 
+	"github.com/moby/moby/api/types/container"
+	"github.com/moby/moby/api/types/mount"
 	"github.com/stretchr/testify/require"
+	"github.com/testcontainers/testcontainers-go"
+	tcexec "github.com/testcontainers/testcontainers-go/exec"
+	"github.com/testcontainers/testcontainers-go/wait"
 	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/component/componentstatus"
 	"go.uber.org/zap/zaptest"
 )
 
-// fakeHost is a component.Host that records fatal-error events reported via
-// componentstatus.ReportStatus, so tests can assert the SIGHUP -> fatal path
-// without wiring the full collector service.
-type fakeHost struct {
-	mu     sync.Mutex
-	fatals []error
-}
+// nopHost is a minimal component.Host for unit tests. The sdnotify extension
+// only stores the host and never calls into it (SIGHUP no longer reports a
+// fatal status event -- otelcol.Collector owns SIGHUP-triggered reload), so
+// GetExtensions returning nil is sufficient.
+type nopHost struct{}
 
-func (h *fakeHost) GetExtensions() map[component.ID]component.Component { return nil }
-
-func (h *fakeHost) Report(e *componentstatus.Event) {
-	if e.Status() == componentstatus.StatusFatalError {
-		h.mu.Lock()
-		h.fatals = append(h.fatals, e.Err())
-		h.mu.Unlock()
-	}
-}
-
-func (h *fakeHost) fatalCount() int {
-	h.mu.Lock()
-	defer h.mu.Unlock()
-	return len(h.fatals)
-}
+func (nopHost) GetExtensions() map[component.ID]component.Component { return nil }
 
 // startFakeNotifySocket opens a datagram socket in a temp dir, points
 // $NOTIFY_SOCKET at it, and returns a channel that receives every payload
@@ -90,7 +79,7 @@ func TestStart_NoNotifySocket_IsNoop(t *testing.T) {
 	t.Setenv("NOTIFY_SOCKET", "")
 
 	s := newSDNotify(&Config{}, zaptest.NewLogger(t))
-	require.NoError(t, s.Start(context.Background(), &fakeHost{}))
+	require.NoError(t, s.Start(context.Background(), nopHost{}))
 	require.True(t, s.isNoop, "expected extension to run as no-op without NOTIFY_SOCKET")
 
 	// Shutdown must be safe even though we never registered a signal handler.
@@ -101,7 +90,7 @@ func TestReady_SendsREADY(t *testing.T) {
 	msgs := startFakeNotifySocket(t)
 
 	s := newSDNotify(&Config{}, zaptest.NewLogger(t))
-	require.NoError(t, s.Start(context.Background(), &fakeHost{}))
+	require.NoError(t, s.Start(context.Background(), nopHost{}))
 	t.Cleanup(func() { _ = s.Shutdown(context.Background()) })
 
 	require.NoError(t, s.Ready())
@@ -117,7 +106,7 @@ func TestNotReady_SendsSTOPPING(t *testing.T) {
 	msgs := startFakeNotifySocket(t)
 
 	s := newSDNotify(&Config{}, zaptest.NewLogger(t))
-	require.NoError(t, s.Start(context.Background(), &fakeHost{}))
+	require.NoError(t, s.Start(context.Background(), nopHost{}))
 	t.Cleanup(func() { _ = s.Shutdown(context.Background()) })
 
 	require.NoError(t, s.NotReady())
@@ -129,32 +118,306 @@ func TestNotReady_SendsSTOPPING(t *testing.T) {
 	}
 }
 
-func TestSIGHUP_ReportsFatalError(t *testing.T) {
-	_ = startFakeNotifySocket(t) // NOTIFY_SOCKET must be set for the signal handler to run
+func TestSIGHUP_SendsRELOADING(t *testing.T) {
+	msgs := startFakeNotifySocket(t) // NOTIFY_SOCKET must be set for the signal handler to run
 
-	host := &fakeHost{}
 	s := newSDNotify(&Config{}, zaptest.NewLogger(t))
-	require.NoError(t, s.Start(context.Background(), host))
+	require.NoError(t, s.Start(context.Background(), nopHost{}))
 	t.Cleanup(func() { _ = s.Shutdown(context.Background()) })
 
 	// Deliver SIGHUP to ourselves; signal.Notify in Start() intercepts it.
 	require.NoError(t, syscall.Kill(os.Getpid(), syscall.SIGHUP))
 
-	waitFor(t, 2*time.Second, func() bool { return host.fatalCount() == 1 },
-		"fatal-error event from SIGHUP handler")
-
-	host.mu.Lock()
-	require.ErrorIs(t, host.fatals[0], errSIGHUP)
-	host.mu.Unlock()
+	select {
+	case got := <-msgs:
+		// RELOADING=1 must be sent together with MONOTONIC_USEC per sd_notify(3).
+		require.Contains(t, got, "RELOADING=1")
+		require.Contains(t, got, "MONOTONIC_USEC=")
+	case <-time.After(2 * time.Second):
+		t.Fatal("no datagram received on fake NOTIFY_SOCKET after SIGHUP")
+	}
 }
 
 func TestShutdown_IsIdempotent(t *testing.T) {
 	_ = startFakeNotifySocket(t)
 
 	s := newSDNotify(&Config{}, zaptest.NewLogger(t))
-	require.NoError(t, s.Start(context.Background(), &fakeHost{}))
+	require.NoError(t, s.Start(context.Background(), nopHost{}))
 
 	require.NoError(t, s.Shutdown(context.Background()))
 	// Second call must not panic on close(closed channel).
 	require.NoError(t, s.Shutdown(context.Background()))
+}
+
+// execAndCollect runs argv in the container and returns combined output.
+// Errors fail the test rather than being returned.
+func execAndCollect(ctx context.Context, t *testing.T, ctr testcontainers.Container, argv ...string) string {
+	t.Helper()
+
+	_, r, err := ctr.Exec(ctx, argv, tcexec.Multiplexed())
+	require.NoError(t, err, "exec failed: %v", argv)
+
+	out, err := io.ReadAll(r)
+	require.NoError(t, err, "reading exec output failed: %v", argv)
+	t.Logf("$ %s\n%s", strings.Join(argv, " "), out)
+
+	return string(out)
+}
+
+// startSystemdContainer builds (or reuses) the sdnotify test image and starts
+// a fresh container running systemd as PID 1. The image is shared across test
+// cases via the Repo/Tag stable name and KeepImage=true, so only the first
+// test in a `go test` invocation pays the ocb build cost; subsequent cases
+// hit Docker's image cache and add only container-start latency.
+//
+// The wait strategy defers to the caller: waitCmd is passed straight to
+// wait.ForExec, so each case decides what "ready" means (unit active vs
+// unit failed vs systemd booted).
+func startSystemdContainer(
+	ctx context.Context,
+	t *testing.T,
+	waitCmd []string,
+	waitExitCode int,
+) testcontainers.Container {
+	t.Helper()
+
+	req := testcontainers.ContainerRequest{
+		FromDockerfile: testcontainers.FromDockerfile{
+			Context:       "../..",
+			Dockerfile:    "extension/sdnotify/testdata/Dockerfile",
+			Repo:          "otelcol-sdnotify-test",
+			Tag:           "latest",
+			KeepImage:     true,
+			PrintBuildLog: true,
+		},
+		HostConfigModifier: func(hc *container.HostConfig) {
+			hc.CapAdd = []string{"SYS_ADMIN", "MKNOD"}
+			hc.SecurityOpt = []string{"seccomp=unconfined"}
+			hc.CgroupnsMode = container.CgroupnsModeHost
+			hc.Mounts = []mount.Mount{
+				{Type: mount.TypeBind, Source: "/sys/fs/cgroup", Target: "/sys/fs/cgroup"},
+			}
+		},
+		WaitingFor: wait.ForExec(waitCmd).
+			WithStartupTimeout(3 * time.Minute).
+			WithPollInterval(2 * time.Second).
+			WithExitCode(waitExitCode),
+	}
+
+	ctr, err := testcontainers.GenericContainer(ctx, testcontainers.GenericContainerRequest{
+		ContainerRequest: req,
+		Started:          true,
+	})
+	require.NoError(t, err, "container failed to start")
+	t.Cleanup(func() { _ = ctr.Terminate(context.Background()) })
+
+	return ctr
+}
+
+// installConfigFile copies a collector config into the running container.
+func installConfigFile(ctx context.Context, t *testing.T, ctr testcontainers.Container, srcHostPath string) {
+	t.Helper()
+
+	err := ctr.CopyFileToContainer(ctx, srcHostPath, "/etc/otelcol/config.yaml", 0o644)
+	require.NoError(t, err, "copying collector config")
+}
+
+// waitForUnitState polls `systemctl show` until ActiveState matches `want`
+// or the timeout expires. Returns the last observed `systemctl show` output
+// so callers can log or assert on it.
+func waitForUnitState(ctx context.Context, t *testing.T, ctr testcontainers.Container, want string, timeout time.Duration) string {
+	t.Helper()
+
+	deadline := time.Now().Add(timeout)
+	var last string
+	for time.Now().Before(deadline) {
+		last = execAndCollect(ctx, t, ctr,
+			"systemctl", "show", "otelcol.service",
+			"-p", "ActiveState", "-p", "SubState", "-p", "Result", "-p", "MainPID", "-p", "NRestarts",
+		)
+		if strings.Contains(last, "ActiveState="+want) {
+			return last
+		}
+		time.Sleep(500 * time.Millisecond)
+	}
+	t.Fatalf("timed out waiting for ActiveState=%s; last show output:\n%s", want, last)
+	return last
+}
+
+// TestSDNotify_SystemdIntegration runs otelcol as a Type=notify unit, and
+// asserts via systemctl that the unit reached active/running - which is only
+// possible if extension.Ready() successfully sent READY=1 over $NOTIFY_SOCKET.
+func TestSDNotify_SystemdIntegration(t *testing.T) {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
+	t.Cleanup(cancel)
+
+	req := testcontainers.ContainerRequest{
+		FromDockerfile: testcontainers.FromDockerfile{
+			Context:       "../..",
+			Dockerfile:    "extension/sdnotify/testdata/Dockerfile",
+			KeepImage:     false,
+			PrintBuildLog: true,
+		},
+		HostConfigModifier: func(hc *container.HostConfig) {
+			// Both kernel capabilities are required to run systemd in a container.
+			// https://github.com/systemd/systemd/blob/main/docs/CONTAINER_INTERFACE.md#what-you-shouldnt-do
+			hc.CapAdd = []string{"SYS_ADMIN", "MKNOD"}
+			// Should run without the default seccomp profile, because it denies mounting.
+			// https://docs.docker.com/engine/security/seccomp
+			hc.SecurityOpt = []string{"seccomp=unconfined"}
+			// Either pre-mount all cgroup hierarchies into the container, or leave that to systemd which will do so if they are missing.
+			// https://github.com/systemd/systemd/blob/main/docs/CONTAINER_INTERFACE.md#execution-environment
+			hc.CgroupnsMode = container.CgroupnsModeHost
+			hc.Mounts = []mount.Mount{
+				{
+					Type:   mount.TypeBind,
+					Source: "/sys/fs/cgroup",
+					Target: "/sys/fs/cgroup",
+				},
+			}
+		},
+		// Wait for systemd to finish booting AND for the otelcol.service to reach `active`.
+		WaitingFor: wait.ForExec([]string{
+			"systemctl", "is-active", "otelcol.service",
+		}).WithStartupTimeout(3 * time.Minute).
+			WithPollInterval(2 * time.Second).
+			WithExitCode(0),
+	}
+
+	ctr, err := testcontainers.GenericContainer(ctx, testcontainers.GenericContainerRequest{
+		ContainerRequest: req,
+		Started:          true,
+	})
+	require.NoError(t, err, "container failed to start; sdnotify integration prerequisites not met")
+	t.Cleanup(func() {
+		_ = ctr.Terminate(context.Background())
+	})
+
+	// Vaidate that otelcol.service is running & healthy.
+	show := execAndCollect(ctx, t, ctr,
+		"systemctl", "show", "otelcol.service",
+		"-p", "ActiveState", "-p", "SubState", "-p", "Result",
+	)
+	require.Contains(t, show, "ActiveState=active")
+	require.Contains(t, show, "SubState=running")
+	require.Contains(t, show, "Result=success")
+
+	// Confirm extension.Ready()'s log line landed in the journal.
+	journal := execAndCollect(ctx, t, ctr,
+		"journalctl", "-u", "otelcol.service", "--no-pager",
+	)
+	require.Contains(t, journal, "sent READY=1 to systemd",
+		"expected sdnotify READY=1 log line in journal, got:\n%s", journal,
+	)
+
+	// Stop the unit and verify a clean exit.
+	_ = execAndCollect(ctx, t, ctr, "systemctl", "stop", "otelcol.service")
+	stopped := execAndCollect(ctx, t, ctr,
+		"systemctl", "show", "otelcol.service",
+		"-p", "ActiveState", "-p", "Result",
+	)
+	require.Contains(t, stopped, "ActiveState=inactive")
+	require.Contains(t, stopped, "Result=success")
+}
+
+// TestSDNotify_SIGHUP_TriggersInProcessReload asserts the SIGHUP contract
+// end-to-end: sending SIGHUP to the collector should
+//   - cause the extension to emit RELOADING=1 (with MONOTONIC_USEC) to systemd,
+//   - trigger otelcol's own in-process reload (service.Shutdown +
+//     setupConfigurationComponents), which drives PipelineWatcher.NotReady and
+//     Ready hooks and produces a fresh STOPPING=1 -> READY=1 pair in the journal,
+//   - NOT cycle the process: MainPID must stay the same, NRestarts stays 0,
+//     and the unit stays ActiveState=active.
+//
+// This is the Type=notify-reload contract: RELOADING=1 -> STOPPING=1 -> READY=1
+// within a single PID.
+func TestSDNotify_SIGHUP_TriggersInProcessReload(t *testing.T) {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
+	t.Cleanup(cancel)
+
+	ctr := startSystemdContainer(ctx, t,
+		[]string{"systemctl", "is-active", "otelcol.service"}, 0)
+
+	beforePID := strings.TrimSpace(
+		execAndCollect(ctx, t, ctr, "systemctl", "show", "otelcol.service", "-p", "MainPID"),
+	)
+
+	_ = execAndCollect(ctx, t, ctr, "systemctl", "kill", "-s", "SIGHUP", "otelcol.service")
+
+	// Wait long enough for otelcol's reload (Shutdown + rebuild) to complete
+	// and for the extension to have emitted RELOADING=1.
+	waitFor(t, 15*time.Second, func() bool {
+		journal := execAndCollect(ctx, t, ctr, "journalctl", "-u", "otelcol.service", "--no-pager")
+		return strings.Contains(journal, "sent RELOADING=1 to systemd") &&
+			strings.Count(journal, "sent READY=1 to systemd") >= 2
+	}, "RELOADING=1 + second READY=1 in journal after SIGHUP-triggered reload")
+
+	// Unit must still be active with the same MainPID: this was an in-process
+	// reload, not a supervisor-driven restart.
+	show := execAndCollect(ctx, t, ctr,
+		"systemctl", "show", "otelcol.service",
+		"-p", "ActiveState", "-p", "SubState", "-p", "MainPID", "-p", "NRestarts",
+	)
+	require.Contains(t, show, "ActiveState=active",
+		"unit should stay active after SIGHUP-triggered in-process reload; show:\n%s", show)
+	require.Contains(t, show, "NRestarts=0",
+		"SIGHUP must not trigger a supervisor restart; show:\n%s", show)
+	afterPID := strings.TrimSpace(
+		execAndCollect(ctx, t, ctr, "systemctl", "show", "otelcol.service", "-p", "MainPID"),
+	)
+	require.Equal(t, beforePID, afterPID,
+		"MainPID must not change on in-process reload; before=%s after=%s", beforePID, afterPID)
+
+	// The journal should show the full reload sequence: RELOADING=1 first,
+	// then STOPPING=1 (from PipelineWatcher.NotReady), then a fresh READY=1.
+	journal := execAndCollect(ctx, t, ctr, "journalctl", "-u", "otelcol.service", "--no-pager")
+	require.Contains(t, journal, "sent RELOADING=1 to systemd",
+		"expected RELOADING=1 log line after SIGHUP; journal:\n%s", journal)
+	require.Contains(t, journal, "sent STOPPING=1 to systemd",
+		"expected STOPPING=1 log line from PipelineWatcher.NotReady during reload; journal:\n%s", journal)
+	// Two READY=1s: the initial boot and the reload.
+	require.GreaterOrEqual(t, strings.Count(journal, "sent READY=1 to systemd"), 2,
+		"expected >=2 READY=1 log lines (boot + reload); journal:\n%s", journal)
+}
+
+// TestSDNotify_StartupFailure_LeavesUnitFailed asserts that when the
+// collector fails during Start() (before Ready() would fire), systemd's
+// Type=notify contract is upheld: the unit lands in failed state rather
+// than hanging on a READY=1 that never comes. This guards against
+// regressions where sdnotify might spuriously send READY=1 too early or
+// swallow the underlying startup error.
+func TestSDNotify_StartupFailure_LeavesUnitFailed(t *testing.T) {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
+	t.Cleanup(cancel)
+
+	// Boot with the default (working) config so systemd itself comes up;
+	// then swap in the broken config and restart the unit. We can't wait
+	// on `is-active` here because the unit will fail -- wait on systemd
+	// itself being up.
+	ctr := startSystemdContainer(ctx, t,
+		[]string{"systemctl", "is-system-running", "--wait"}, 0)
+
+	installConfigFile(ctx, t, ctr, "testdata/otelcol-bad-config.yaml")
+
+	// `systemctl restart` will block until start-limit is reached or the
+	// unit fails; with Restart=no it fails on the first attempt. Ignore
+	// exit code -- we assert on the resulting unit state.
+	_, _, _ = ctr.Exec(ctx, []string{"systemctl", "restart", "otelcol.service"})
+
+	show := waitForUnitState(ctx, t, ctr, "failed", 60*time.Second)
+	require.Contains(t, show, "ActiveState=failed",
+		"unit should be failed after startup with broken config; show:\n%s", show)
+
+	// The extension must NOT have logged a spurious READY=1 -- Ready()
+	// should never have been called on a failed startup.
+	journal := execAndCollect(ctx, t, ctr,
+		"journalctl", "-u", "otelcol.service", "--no-pager",
+	)
+	// Filter to just the failed incarnation (after the config swap).
+	// Simplest check: no successful READY=1 line since the last restart
+	// attempt would be a bug, but the first boot did send one. Assert on
+	// the presence of a collector-side error log instead.
+	require.True(t,
+		strings.Contains(journal, "error") || strings.Contains(journal, "failed"),
+		"expected error/failed log line from broken-config startup; journal:\n%s", journal)
 }
