--- extension/sdnotify/testdata/otelcol-restart.service
@@ -1,15 +1,0 @@
-[Unit]
-Description=OpenTelemetry Collector (sdnotify integration test, restart-on-failure)
-After=network-online.target
-Wants=network-online.target
-
-[Service]
-Type=notify
-NotifyAccess=main
-ExecStart=/bin/otelcol --config=/etc/otelcol/config.yaml
-TimeoutStartSec=60
-Restart=on-failure
-RestartSec=100ms
-
-[Install]
-WantedBy=multi-user.target
